Custom VZbot VZ-Printhead mount
I've created a custom mount for the Euclid Probe for the VZ-Printhead. (Euclid mount VZ-printhead v5)

It needs to be assembled onto the 5015 fan at the back.

The dock that I used is the VoronGantyMountRev_4 filed down a little bit along with the RailCoreTopMountLong, assembled the other way to make the Mount fit.

The image should make the assembly clearer.

I've added the docks from the other repos here to make it easier to find the full assembly.